BUG: sparse: trigger 64-bit index cast in more locations

+++ scipy/sparse/coo.py
@@ -300,7 +300,9 @@ class coo_matrix(_data_matrix, _minmax_mixin):
             data = np.empty(self.nnz, dtype=upcast(self.dtype))
 
             coo_tocsr(N, M, self.nnz,
-                      self.col, self.row, self.data,
+                      self.col.astype(idx_dtype),
+                      self.row.astype(idx_dtype),
+                      self.data,
                       indptr, indices, data)
 
             A = csc_matrix((data, indices, indptr), shape=self.shape)
@@ -339,8 +341,12 @@ class coo_matrix(_data_matrix, _minmax_mixin):
             data = np.empty(self.nnz, dtype=upcast(self.dtype))
 
             coo_tocsr(M, N, self.nnz,
-                      self.row, self.col, self.data,
-                      indptr, indices, data)
+                      self.row.astype(idx_dtype),
+                      self.col.astype(idx_dtype),
+                      self.data,
+                      indptr,
+                      indices,
+                      data)
 
             A = csr_matrix((data, indices, indptr), shape=self.shape)
             A.sum_duplicates()
